EUI-150 moving existing editor config to cong.json (#31)
* adding favicon

* adding editor config to config.js working version

* cleaing some tests

* moving case config to a global place

* renaming the ccd config file

* fixing tests
