+++ src/app/services/ccd-config/ccd-case.config.ts
@@ -3,7 +3,8 @@ import {
   AbstractAppConfig,
   CaseEditorConfig
 } from '@hmcts/ccd-case-ui-toolkit';
-import { environment as mainConfig } from '../environments/environment';
+import {AppConfigService} from '../config/configuration.services';
+
 
 /**
  * see more:
@@ -13,10 +14,11 @@ import { environment as mainConfig } from '../environments/environment';
 
 @Injectable()
 export class AppConfig extends AbstractAppConfig {
-  // TODO perhaps move the config to store or maybe make it part of assets/config.json
-  protected config: CaseEditorConfig = mainConfig.CaseEditorConfig;
-  constructor() {
+
+  protected config: CaseEditorConfig;
+  constructor(private appConfigService: AppConfigService) {
     super();
+    this.config =  this.appConfigService.getEditorConfiguration() || {};
   }
 
   public load(): Promise<void> {
