+++ src/cases/containers/case-create/case-create.component.spec.ts
@@ -17,7 +17,7 @@ import {
   RequestOptionsBuilder,
   SearchFiltersModule,
 } from '@hmcts/ccd-case-ui-toolkit';
-import {AppConfig} from '../../case.config';
+import {AppConfig} from '../../../app/services/ccd-config/ccd-case.config';
 import {ScrollToService} from '@nicky-lenaers/ngx-scroll-to';
 import {RouterTestingModule} from '@angular/router/testing';
 import {HttpClientModule} from '@angular/common/http';
@@ -26,10 +26,12 @@ import {HttpModule} from '@angular/http';
 import {SharedModule} from '../../../app/shared/shared.module';
 import {AppConfigService} from '../../../app/services/config/configuration.services';
 import {CasesCreateComponent} from './case-create.component';
+import {ProvidersModule} from '../../../app/providers/providers.module';
 
 class MockSortService {
   features = {};
-  getFeatureToggle() {}
+  getFeatureToggle() {};
+  getEditorConfiguration() {};
 }
 describe('CaseCaseComponent', () => {
   let component: CasesCreateComponent;
@@ -59,9 +61,10 @@ describe('CaseCaseComponent', () => {
         CaseEditWizardGuard,
         RouterHelperService,
         DocumentManagementService,
-        AppConfig,
         AppConfigService,
         RequestOptionsBuilder,
+        AppConfigService,
+        AppConfig,
         {
           provide: SearchService,
           useValue: {
