+++ src/cases/containers/case-list/creats-list.component.spec.ts
@@ -5,7 +5,7 @@ import {
   HttpErrorService, AbstractAppConfig, CaseEditWizardGuard, RouterHelperService,
   DocumentManagementService, PageValidationService, PlaceholderService, SearchService, RequestOptionsBuilder, SearchFiltersModule
 } from '@hmcts/ccd-case-ui-toolkit';
-import {AppConfig} from '../../case.config';
+import {AppConfig} from '../../../app/services/ccd-config/ccd-case.config';
 import {ScrollToService} from '@nicky-lenaers/ngx-scroll-to';
 import {RouterTestingModule} from '@angular/router/testing';
 import {HttpClientModule} from '@angular/common/http';
