+++ src/cases/containers/case-search/case.search.component.spec.ts
@@ -17,7 +17,7 @@ import {
   RequestOptionsBuilder,
   SearchFiltersModule,
 } from '@hmcts/ccd-case-ui-toolkit';
-import {AppConfig} from '../../case.config';
+import {AppConfig} from '../../../app/services/ccd-config/ccd-case.config';
 import {ScrollToService} from '@nicky-lenaers/ngx-scroll-to';
 import {RouterTestingModule} from '@angular/router/testing';
 import {HttpClientModule} from '@angular/common/http';
